Clean utilities, moving ROOT plotting functions to utilities_plot
