+++ Analyses/ALICE_D2H_vs_mult_pp13/plot_final_results_lc.py
@@ -21,12 +21,12 @@ import yaml
 from ROOT import TFile, gStyle, gROOT, TH1F, TGraphAsymmErrors, TH1
 from ROOT import kBlue, kAzure, kOrange, kGreen, kBlack, kRed, kWhite
 from ROOT import Double
-from machine_learning_hep.utilities import plot_histograms, save_histograms, Errors
-from machine_learning_hep.utilities import calc_systematic_multovermb
-from machine_learning_hep.utilities import divide_all_by_first_multovermb
-from machine_learning_hep.utilities import divide_by_eachother, divide_by_eachother_barlow
-from machine_learning_hep.utilities import calc_systematic_mesonratio
-from machine_learning_hep.utilities import calc_systematic_mesondoubleratio
+from machine_learning_hep.utilities_plot import plot_histograms, save_histograms, Errors
+from machine_learning_hep.utilities_plot import calc_systematic_multovermb
+from machine_learning_hep.utilities_plot import divide_all_by_first_multovermb
+from machine_learning_hep.utilities_plot import divide_by_eachother, divide_by_eachother_barlow
+from machine_learning_hep.utilities_plot import calc_systematic_mesonratio
+from machine_learning_hep.utilities_plot import calc_systematic_mesondoubleratio
 
 def results(histos_central, systematics, title, legend_titles, x_label, y_label,
             save_path, ratio, **kwargs):
