+++ Analyses/ALICE_D2H_vs_mult_pp13/plot_ml_std_comp.py
@@ -20,7 +20,7 @@ from math import sqrt
 # pylint: disable=import-error, no-name-in-module, unused-import
 import yaml
 from ROOT import TFile, gStyle, gROOT
-from machine_learning_hep.utilities import plot_histograms
+from machine_learning_hep.utilities_plot import plot_histograms
 
 FILES_NOT_FOUND = []
