+++ machine_learning_hep/plot_hfptspectrum.py
@@ -26,7 +26,7 @@ from ROOT import TFile, TH1F, TCanvas
 from ROOT import gStyle, TLegend, TLatex
 from ROOT import gROOT, kRed, kGreen, kBlack, kBlue, kOrange, kViolet, kAzure
 from ROOT import TStyle, gPad
-from machine_learning_hep.utilities import plot_histograms
+from machine_learning_hep.utilities_plot import plot_histograms
 
 FILES_NOT_FOUND = []
 # One single particle ratio
