+++ machine_learning_hep/plot_hfptspectrum_years.py
@@ -19,7 +19,7 @@ import os
 # pylint: disable=import-error, no-name-in-module, unused-import
 import yaml
 from ROOT import gROOT, TFile
-from machine_learning_hep.utilities import plot_histograms
+from machine_learning_hep.utilities_plot import plot_histograms
 
 FILES_NOT_FOUND = []
