+++ machine_learning_hep/processer.py
@@ -31,8 +31,9 @@ from machine_learning_hep.bitwise import filter_bit_df, tag_bit_df
 from machine_learning_hep.utilities import selectdfquery, selectdfrunlist, merge_method
 from machine_learning_hep.utilities import list_folders, createlist, appendmainfoldertolist
 from machine_learning_hep.utilities import create_folder_struc, seldf_singlevar, openfile
-from machine_learning_hep.utilities import mergerootfiles, z_calc, z_gen_calc, scatterplot
+from machine_learning_hep.utilities import mergerootfiles, z_calc, z_gen_calc
 from machine_learning_hep.utilities import get_timestamp_string
+from machine_learning_hep.utilities_plot import scatterplotroot
 from machine_learning_hep.models import apply # pylint: disable=import-error
 #from machine_learning_hep.globalfitter import fitter
 from machine_learning_hep.selectionutils import getnormforselevt
@@ -784,11 +785,11 @@ class Processer: # pylint: disable=too-many-instance-attributes
         minrvar = [0, 0, 0]
         maxrvar = [1500, 200, .5]
         fileevtroot = TFile.Open(self.l_evtvalroot[file_index], "recreate")
-        hv0mvsperc = scatterplot(dfevt, "perc_v0m", "v0m_corr", 50000, 0, 100, 200, 0., 2000.)
+        hv0mvsperc = scatterplotroot(dfevt, "perc_v0m", "v0m_corr", 50000, 0, 100, 200, 0., 2000.)
         hv0mvsperc.SetName("hv0mvsperc")
         hv0mvsperc.Write()
         dfevtnorm = pickle.load(openfile(self.l_evtorig[file_index], "rb"))
-        hntrklsperc = scatterplot(dfevt, "perc_v0m", "n_tracklets_corr", 50000, 0, 100, 200, 0., 2000.)
+        hntrklsperc = scatterplotroot(dfevt, "perc_v0m", "n_tracklets_corr", 50000, 0, 100, 200, 0., 2000.)
         hntrklsperc.SetName("hntrklsperc")
         hntrklsperc.Write()
         for ivar, var in enumerate(varlist):
